Add Admin menu generic class
